miku mode attempt #2
The first version was compiled on a computer with a different WiFi library, which caused every client to think it's MAC was 00:00:00:00:00:00, causing them to all have the same client ID and rapidly disconnect and reconnect to the MQTT broker. This version was compiled on a different computer which should have the "working" version of the library.
"Working" is in quotation marks, because I think WiFi.macAddress() is not supposed to work before WiFi.Begin(), yet here it does.